requirements.txt
pandas
numpy
scikit-learn
flask
seaborn
matplotlib
xgboost
catboost
dill
gunicorn
ipykernel
python-dotenv
pymongo
certifi
pymongo[srv]
pyaml
mlflow
dagshub
fastapi
uvicorn
python-multipart